+++ apps/payment/views.py
@@ -48,19 +48,19 @@ def payment(request, ticket_type_id):
         return error_response
 
     if ticket_type.lan.has_ticket(request.user) or request.user in ticket_type.lan.paid_attendees:
-        messages.info(request, _(u'You already have a ticket for this LAN.'))
+        messages.warning(request, _(u'You already have a ticket for this LAN.'))
         return error_response
 
     if not ticket_type.is_available():
-        messages.info(request, _(u'This ticket is not yet available.'))
+        messages.warning(request, _(u'This ticket is not yet available.'))
         return error_response
 
     if ticket_type.is_sold_out():
-        messages.info(request, _(u'All tickets have sold out.'))
+        messages.warning(request, _(u'All tickets have sold out.'))
         return error_response
 
     if request.user not in ticket_type.lan.attendees:
-        messages.info(request, _(u'You must attend first to buy a ticket for this LAN.'))
+        messages.error(request, _(u'You must attend first to buy a ticket for this LAN.'))
         return error_response
 
     if request.method == 'POST':